+++ apps/api-sdk/src/logs/types.ts
@@ -5,12 +5,20 @@ export enum Queues {
   logs = "logs",
 }
 
-export type QLogsInput = {
-  envId: string;
-  datapointRunId: string | undefined;
-  trace: TraceInput | undefined;
-  log: LogInput;
-};
+export type QLogsInput =
+  | {
+      action: "create";
+      envId: string;
+      datapointRunId: string | undefined;
+      trace: TraceInput | undefined;
+      log: LogInput;
+    }
+  | {
+      action: "end";
+      envId: string;
+      logId: string;
+      payload: Pick<LogInput, "output" | "endTime" | "extra">;
+    };
 
 const traceWithLogs = Prisma.validator<Prisma.TraceDefaultArgs>()({
   include: {