+++ apps/montelo/src/Montelo.ts
@@ -1,4 +1,5 @@
 import * as cuid from "@paralleldrive/cuid2";
+import { Log } from "./Log";
 import { MonteloClient } from "./MonteloClient";
 import { LogInput, LogInputSourceEnum } from "./client";
 import { MonteloDatapoints, MonteloDatasets, MonteloExperiments } from "./core";
@@ -32,11 +33,15 @@ export class Montelo {
     this.experiments = new MonteloExperiments(this.monteloClient);
   }
 
-  public log(log: LogParams) {
+  public async log(log: LogParams) {
+    const id = cuid.createId();
     const now = new Date().toISOString();
     const startTime = log.startTime || now;
     const endTime = log.endTime || now;
-    void this.monteloClient.createLog({ ...log, startTime, endTime, source: LogInputSourceEnum.Manual });
+    await this.monteloClient.createLog({ ...log, id, startTime, endTime, source: LogInputSourceEnum.Manual });
+
+    const logInstance = new Log({ id, monteloClient: this.monteloClient });
+    return logInstance;
   }
 
   public trace(trace: TraceParams): Trace {