+++ apps/api-sdk/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @montelo/api-sdk
 
+## 1.1.67
+
+### Patch Changes
+
+- Crews & Logs End Method
+- Updated dependencies
+  - @montelo/api-common@1.1.67
+  - @montelo/db@1.1.67
+
 ## 1.1.66
 
 ### Patch Changes