+++ apps/crews/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @montelo/crews
 
+## 1.1.67
+
+### Patch Changes
+
+- Crews & Logs End Method
+- Updated dependencies
+  - montelo@1.1.67
+
 ## 1.1.66
 
 ### Patch Changes